requirements.txt
PyYAML==3.12
sh==1.12.14
